docs/module-04-vla-multimodal/solutions/chapter-03/requirements.txt
# Chapter 3: Cognitive Planning with LLMs - Python Dependencies

# ROS 2 Python client library
rclpy>=3.0.0

# OpenAI API for LLM planning
openai>=1.0.0

# Environment variable management
python-dotenv>=1.0.0

# ROS 2 message types (installed with ROS 2)
# geometry_msgs
# std_msgs
# action_msgs

# Optional: Nav2 messages for navigation integration
# nav2_msgs (install via apt: ros-humble-nav2-msgs)

# Note: ROS 2 Humble must be installed separately
# Installation: https://docs.ros.org/en/humble/Installation.html
